requirements.txt
fusepy==2.0.4
PyYAML==3.12
